feature/basic-crud: cria apis para get, put e delete.

+++ infrastructure/src/main/kotlin/com/school/jpaRepository/StudentJpaRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.stereotype.Repository
 import org.springframework.data.jpa.repository.JpaRepository
 
 @Repository
-interface StudentJpaRepository : JpaRepository<Student, Long>
+interface StudentJpaRepository : JpaRepository<Student, Int>
